+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## 0.0.6
+- Files are located in a sub-folder of the temp directory
+- Tempcache accepts a `name` parameter as name of is sub-folder
 - Added `source` argument to `TempCache` to further differentiate caches
 - Using `inspect` module `BoundArguments` to create task digests